Implementa Swagger y documentacion de ms-sucursales

+++ ms_sucursales/src/main/java/com/arriendo/ms_sucursales/service/SucursalService.java
@@ -28,23 +28,30 @@ public class SucursalService {
     private RegionRepository regionRepository;
 
     public List<Sucursal> obtenerTodas() {
+
         log.info("Listando todas las sucursales");
+
         return sucursalRepository.findAll();
     }
 
     public List<Sucursal> buscarOperativas() {
+
         log.info("Buscando sucursales operativas");
+
         return sucursalRepository.buscarSucursalesOperativas();
     }
 
-    public Sucursal obtenerPorId(Integer id) {
+    public Sucursal obtenerPorId(Long id) {
+
         log.info("Buscando sucursal con id {}", id);
+
         return sucursalRepository.findById(id)
                 .orElseThrow(() ->
                         new ResourceNotFoundException("Sucursal no encontrada"));
     }
 
     public SucursalResponseDTO guardar(SucursalRequestDTO dto) {
+
         log.info("Guardando nueva sucursal");
 
         Region region = regionRepository.findById(dto.getRegionId())
@@ -52,12 +59,14 @@ public class SucursalService {
                         new ResourceNotFoundException("Región no encontrada"));
 
         Sucursal sucursal = SucursalMapper.toEntity(dto, region);
+
         Sucursal sucursalGuardada = sucursalRepository.save(sucursal);
 
         return SucursalMapper.toDTO(sucursalGuardada);
     }
 
-    public Sucursal actualizar(Integer id, SucursalRequestDTO dto) {
+    public Sucursal actualizar(Long id, SucursalRequestDTO dto) {
+
         log.info("Actualizando sucursal con id {}", id);
 
         Sucursal sucursal = obtenerPorId(id);
@@ -76,10 +85,12 @@ public class SucursalService {
         return sucursalRepository.save(sucursal);
     }
 
-    public void eliminar(Integer id) {
+    public void eliminar(Long id) {
+
         log.info("Eliminando sucursal con id {}", id);
 
         Sucursal sucursal = obtenerPorId(id);
+
         sucursalRepository.delete(sucursal);
     }
 }
